requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
pdfplumber==0.10.3 